+++ src/BoardManager.cpp
@@ -4,7 +4,6 @@ BoardManager::BoardManager(shared_ptr<EventManager> eventManager, shared_ptr<Tex
   this -> textLoader = textLoader;
   this -> mapFactory = unique_ptr<MapFactory>(new MapFactory(new MapChoices(1,1,10)));
   this -> eventManager = eventManager;
-  this -> newMap();
   this -> registerDelegates();
 
 }
@@ -74,21 +73,59 @@ void BoardManager::handleTowerCreation(const EventInterface& event){
   //the string ID identifying the towers
   string towerTypeID = tcEventData -> towerTypeID;
 
+  //get the numeric identifier for the tower type
+  int towerTypeNum = textLoader -> getConstant(string("IDS_"+towerTypeID));
 
+  //decode the ID so that it is a row and column
+  int row = towerPosID / getYDim();
+  int col = towerPosID % getXDim();
 
+  //place the tower value on the above floor grid
+  floorGrid.at(row).at(col) = towerTypeNum;
 }
 
+/*
+ * @return true if no map is stored
+ */
+bool BoardManager::hasMap(){
+  return floorGrid.size()!=0 || distances.size()!=0 || aboveFloorGrid.size()!=0;
+}
 /*
  * @return true if the selected grid has a tower or an obstacle
  */
 bool BoardManager::isTowerOrObstacle(int row, int col){
+  assert(hasMap()== true);
   return aboveFloorGrid.at(row).at(col) > 0 || aboveFloorGrid.at(row).at(col) < -1;
 }
+/*
+ * @return true if the selected grid has an exit
+ */
+bool BoardManager::isExit(int row, int col){
+  assert(hasMap()== true);
+  return aboveFloorGrid.at(row).at(col) == 0;
+}
+/*
+ * @return true if the selected grid has a path
+ */
+bool BoardManager::isPath(int row, int col){
+  assert(hasMap()== true);
+  return floorGrid.at(row).at(col) > 0 ;
+}
+/*
+ * @return true if the selected grid has a path
+ */
+bool BoardManager::isObstacle(int row, int col){
+  assert(hasMap()== true);
+  return aboveFloorGrid.at(row).at(col) < -1 ;
+}
+
 
 int BoardManager::getYDim(){
+  assert(hasMap()== true);
   return mapFactory -> getYDim();
 }
 int BoardManager::getXDim(){
+  assert(hasMap()== true);
   return mapFactory -> getXDim();
 }
 void BoardManager::setMapCustomizationChoices(MapChoices * newCustomization){
@@ -105,12 +142,15 @@ void BoardManager::setMapEntryChoice(int pathEntryChoice){
 }
 
 vector<vector<int>>const & BoardManager::getDistances(){
+  assert(hasMap()== true);
   return distances;
 }
 vector<int>const & BoardManager::getDistancesRow(int row){
+  assert(hasMap()== true);
   return distances.at(row);
 }
 vector<int>const & BoardManager::getDistanceCol(int col){
+  assert(hasMap()== true);
   static vector<int> distanceCol;
   distanceCol.clear();
 
@@ -121,16 +161,20 @@ vector<int>const & BoardManager::getDistanceCol(int col){
 }
 
 int BoardManager::getDistanceRowCol(int row, int col){
+  assert(hasMap()== true);
   return distances.at(row).at(col);
 }
 
 vector<vector<int>>const & BoardManager::getAboveFloor(){
+  assert(hasMap()== true);
   return aboveFloorGrid;
 }
 vector<int>const & BoardManager::getAboveFloorRow(int row){
+  assert(hasMap()== true);
   return aboveFloorGrid.at(row);
 }
 vector<int>const & BoardManager::getAboveFloorCol(int col){
+  assert(hasMap()== true);
   static vector<int> aboveFloorCol;
   aboveFloorCol.clear();
 
@@ -140,16 +184,20 @@ vector<int>const & BoardManager::getAboveFloorCol(int col){
   return aboveFloorCol;
 }
 int BoardManager::getAboveFloorRowCol(int row, int col){
+  assert(hasMap()== true);
   return aboveFloorGrid.at(row).at(col);
 }
 
 vector<vector<int>>const & BoardManager::getFloor(){
+  assert(hasMap()== true);
   return floorGrid;
 }
 vector<int>const & BoardManager::getFloorRow(int row){
+  assert(hasMap()== true);
   return floorGrid.at(row);
 }
 vector<int>const & BoardManager::getFloorCol(int col){
+  assert(hasMap()== true);
   static vector<int> floorCol;
   floorCol.clear();
 
@@ -159,9 +207,27 @@ vector<int>const & BoardManager::getFloorCol(int col){
   return floorCol;
 }
 int BoardManager::getFloorRowCol(int row, int col){
+  assert(hasMap()== true);
   return floorGrid.at(row).at(col);
 }
 
-vector<int> BoardManager::getEntryPositions(){
+vector<int>& BoardManager::getEntryPositions(){
+  assert(hasMap()== true);
   return mapFactory->getEntryPos();
+}
+
+unordered_map<int,pair<int,int>>& BoardManager::getAllObstacles(){
+  assert(hasMap()== true);
+  if(allObstacles.size() != 0){
+    return allObstacles;
+  }
+  for(int i=0; i<getYDim(); i++){
+    for(int j=0; j<getXDim(); j++){
+      if(aboveFloorGrid.at(i).at(j) < -2){
+        intPair pos (i,j);
+        allObstacles.insert({aboveFloorGrid.at(i).at(j), pos});
+      }
+    }
+  }
+  return allObstacles;
 }