+++ src/Obstacles/CafeteriaTable.cpp
@@ -1,7 +1,7 @@
 #include "../include/Obstacles/CafeteriaTable.hpp"
 
-CafeteriaTable::CafeteriaTable(int row, int col, int price){
-  this -> row = row;
-  this -> col = col;
-  this -> price = price;
+CafeteriaTable::CafeteriaTable(shared_ptr<TextLoader> textLoader, string towerTypeID) : Obstacle(){
+  this -> towerTypeID = towerTypeID;
+  this -> textLoader = textLoader;
+  this -> price = textLoader -> getConstant(string("IDS_CT_PR"));
 }