+++ src/TowerManager.cpp
@@ -3,16 +3,22 @@
 /*
  * Constructor for a tower manager
  * @param eventManager: the event manager class that directs events
- * @param yDim: the y dimension of the grid (rows)
- * @param xDim: the x dimension of the grid (cols)
  */
-TowerManager::TowerManager(shared_ptr<EventManager> eventManager, shared_ptr<TextLoader> textLoader, int yDim, int xDim){
+TowerManager::TowerManager(shared_ptr<EventManager> eventManager, shared_ptr<TextLoader> textLoader){
   this -> eventManager = eventManager;
   this -> textLoader = textLoader;
+  this -> registerDelegates();
+  this -> populateObstacles();
+  this -> populateTowersToChoose();
+}
+
+/*
+ * Set the x dimension (cols)
+ * and the y dimension (rows)
+ */
+void TowerManager::setDimensions(int xDim, int yDim){
   this -> xDim = xDim;
   this -> yDim = yDim;
-  this -> registerDelegates();
-  this -> populateTowersToChoose();
 }
 
 /*
@@ -33,7 +39,38 @@ void TowerManager::registerDelegates(){
 }
 
 /*
- * Add all the basic towers to the vector of possible towers
+ * Add all obstacles to the vector of possible towers
+ */
+void TowerManager::populateObstacles(){
+  //get the identifier for each type of obstacle
+  string cafeteriaTableID = textLoader ->  getString("IDS_CTO");
+  string trashCanID = textLoader -> getString("IDS_TCO");
+
+  //make a pointer to each type of obstacle
+  shared_ptr<Obstacle> cafeteriaTable = make_shared<CafeteriaTable>(textLoader, cafeteriaTableID);
+  shared_ptr<Obstacle> trashCan = make_shared<TrashCan>(textLoader, trashCanID);
+
+  //all the obstacles serve as their own upgrade
+  //used to give obstacles compatability in situations where
+  //you want what can be bought at a tile and it can be nothing, a tower, or an obstacle
+  vector<shared_ptr<TowerInterface>> allCafeteriaTableUpgrades;
+  allCafeteriaTableUpgrades.push_back(cafeteriaTable);
+  vector<shared_ptr<TowerInterface>> allTrashCanUpgrades;
+  allTrashCanUpgrades.push_back(trashCan);
+
+  //add the towers to tower upgrades
+  allTowerUpgrades.insert({cafeteriaTableID, allCafeteriaTableUpgrades});
+  allTowerUpgrades.insert({trashCanID, allTrashCanUpgrades});
+
+  //add the obstacles  to towers types
+  allTowerTypes.insert({cafeteriaTableID, cafeteriaTable});
+  allTowerTypes.insert({trashCanID, trashCan});
+}
+
+/*
+ * Add all the basic towers to the map of possible tower upgrades
+ * keyed on the not a tower. and add them to the list of
+ * placeable towers
  */
 void TowerManager::populateTowersToChoose(){
 
@@ -63,14 +100,22 @@ void TowerManager::populateTowersToChoose(){
   noTowerAtTileUpgrades.push_back(normalFry);
   noTowerAtTileUpgrades.push_back(miniMMS);
 
-  //add the towers to towers to choose
-  allTowersToChoose.insert({noTowerAtTileID, noTowerAtTileUpgrades});
+  //add the towers to tower upgrades
+  allTowerUpgrades.insert({noTowerAtTileID, noTowerAtTileUpgrades});
+
+  //add the towers  to towers types
+  allTowerTypes.insert({cheesePizzaID, cheesePizza});
+  allTowerTypes.insert({sodaID, soda});
+  allTowerTypes.insert({normalFryID, normalFry});
+  allTowerTypes.insert({miniMMSID, miniMMS});
+
 
   this -> populateTowerUpgrades();
 }
 
 /*
  * Add all the upgrades for each tower to a map of possible upgrades
+ * and add them to the list of placeable towers
  */
 void TowerManager::populateTowerUpgrades(){
   //get the maximum number of melee units allowed per melee tower
@@ -78,6 +123,7 @@ void TowerManager::populateTowerUpgrades(){
 
   populateTowerUpgradesLvl1(maxMeleeUnits);
   populateTowerUpgradesLvl2(maxMeleeUnits);
+  populateTowerUpgradesLvl3(maxMeleeUnits);
 
 }
 /*
@@ -115,11 +161,16 @@ void TowerManager::populateTowerUpgradesLvl1(int maxMeleeUnits){
 
   //insert the vector into the map of all possible upgrades keyed on the string identifier
   //from the type of tower you upgrade from
-  allTowersToChoose.insert({cheesePizzaID, cheesePizzaUpgrades});
-  allTowersToChoose.insert({sodaID, sodaUpgrades});
-  allTowersToChoose.insert({normalFryID, normalFryUpgrades});
-  allTowersToChoose.insert({miniMMSID, miniMMSUpgrades});
+  allTowerUpgrades.insert({cheesePizzaID, cheesePizzaUpgrades});
+  allTowerUpgrades.insert({sodaID, sodaUpgrades});
+  allTowerUpgrades.insert({normalFryID, normalFryUpgrades});
+  allTowerUpgrades.insert({miniMMSID, miniMMSUpgrades});
 
+  //add the towers  to towers types
+  allTowerTypes.insert({pepperoniPizzaID, pepperoniPizza});
+  allTowerTypes.insert({energyDrinkID, energyDrink});
+  allTowerTypes.insert({crinkleFryID, crinkleFry});
+  allTowerTypes.insert({normalMMSID, normalMMS});
 }
 /*
  * Add all the upgrades for each tower to a map of possible upgrades for level 2
@@ -168,25 +219,106 @@ void TowerManager::populateTowerUpgradesLvl2(int maxMeleeUnits){
 
   //insert the vector into the map of all possible upgrades keyed on the string identifier
   //from the type of tower you upgrade from
-  allTowersToChoose.insert({pepperoniPizzaID, pepperoniPizzaUpgrades});
-  allTowersToChoose.insert({crinkleFryID, crinkleFryUpgrades});
-  allTowersToChoose.insert({energyDrinkID, energyDrinkUpgrades});
-  allTowersToChoose.insert({normalMMSID, normalMMSUpgrades});
+  allTowerUpgrades.insert({pepperoniPizzaID, pepperoniPizzaUpgrades});
+  allTowerUpgrades.insert({crinkleFryID, crinkleFryUpgrades});
+  allTowerUpgrades.insert({energyDrinkID, energyDrinkUpgrades});
+  allTowerUpgrades.insert({normalMMSID, normalMMSUpgrades});
+
+  //add the towers  to towers types
+  allTowerTypes.insert({deepDishID, deepDish});
+  allTowerTypes.insert({meatLoversID, meatLovers});
+  allTowerTypes.insert({spicyFryID, spicyFry});
+  allTowerTypes.insert({waffleFryID, waffleFry});
+  allTowerTypes.insert({slushieID, slushie});
+  allTowerTypes.insert({gravyID, gravy});
+  allTowerTypes.insert({peanutButterMMSID, peanutButterMMS});
+  allTowerTypes.insert({peanutMMSID, peanutMMS});
 }
 
 /*
+ * Add all the upgrades for each tower to a map of possible upgrades for level 3
+ * @param maxMeleeUnits: the maximum number of units a melee tower can have
+ */
+void TowerManager::populateTowerUpgradesLvl3(int maxMeleeUnits){
+  //get the identifier for each level 3 type of tower
+  string deepDishID = textLoader -> getString(string("IDS_DDT"));
+  string meatLoversID = textLoader -> getString(string("IDS_MLT"));
+  string spicyFryID = textLoader -> getString(string("IDS_SFT"));
+  string waffleFryID = textLoader -> getString(string("IDS_WFT"));
+  string slushieID = textLoader -> getString(string("IDS_SLT"));
+  string gravyID = textLoader -> getString(string("IDS_GT"));
+  string peanutButterMMSID = textLoader -> getString(string("IDS_PBMMT"));
+  string peanutMMSID = textLoader -> getString(string("IDS_PMMT"));
+
+  //push each tower into a vector of possible upgrades
+  //there are none since this is the last level of towers
+  vector<shared_ptr<TowerInterface>> deepDishUpgrades;
+  vector<shared_ptr<TowerInterface>> meatLoversUpgrades;
+  vector<shared_ptr<TowerInterface>> spicyFryUpgrades;
+  vector<shared_ptr<TowerInterface>> waffleFryUpgrades;
+  vector<shared_ptr<TowerInterface>> slushieUpgrades;
+  vector<shared_ptr<TowerInterface>> gravyUpgrades;
+  vector<shared_ptr<TowerInterface>> peanutButterMMSUpgrades;
+  vector<shared_ptr<TowerInterface>> peanutMMSUpgrades;
+
+  //insert the vector into the map of all possible upgrades keyed on the string identifier
+  //from the type of tower you upgrade from
+  allTowerUpgrades.insert({deepDishID, deepDishUpgrades});
+  allTowerUpgrades.insert({meatLoversID, meatLoversUpgrades});
+  allTowerUpgrades.insert({spicyFryID, spicyFryUpgrades});
+  allTowerUpgrades.insert({waffleFryID, waffleFryUpgrades});
+  allTowerUpgrades.insert({slushieID, slushieUpgrades});
+  allTowerUpgrades.insert({gravyID, gravyUpgrades});
+  allTowerUpgrades.insert({peanutButterMMSID, peanutButterMMSUpgrades});
+  allTowerUpgrades.insert({peanutMMSID, peanutMMSUpgrades});
+}
+
+/*
+ * default version that does not include a specific upgrade decision
  * @return the price of the tower at the row and col
  */
 int TowerManager::getTowerPrice(int row, int col){
-  return getTowerPlaced(row, col) -> getPrice();
+  string towerID;
+
+  //if there is no tower here
+  if(towersPlaced.find(row*xDim+col) == towersPlaced.end()){
+    //get the identifier for no tower type chosen
+    towerID = textLoader -> getString("IDS_NT");
+
+  }else{
+     auto tower = towersPlaced.at(row*xDim+col);
+     towerID = tower -> getType();
+  }
+
+  // get the vector of towers we can upgrade to
+  //grab the first tower in that list
+  //and return its price
+  int price = allTowerUpgrades.at(towerID).at(0) -> getPrice();
+
+  return price;
+}
+
+/*
+ * @return the price of the specific tower
+ */
+int TowerManager::getTowerPrice(string towerTypeID){
+  assert(allTowerTypes.find(towerTypeID) != allTowerTypes.end());
+  return allTowerTypes.at(towerTypeID) -> getPrice();
 }
 
 
 /*
- * @return vector<TowerInterface>: all the towers the player can choose from
+ * @return unordered_map<string, vector<TowerInterface>>: all the towers the player can choose from
  */
 unordered_map<string, vector<shared_ptr<TowerInterface>>>& TowerManager::getAllTowersToChoose(){
-  return allTowersToChoose;
+  return allTowerUpgrades;
+}
+
+/*
+ * @return vector<TowerInterface>: all the upgrades for a partiuclar tower type
+ */
+vector<shared_ptr<TowerInterface>>& TowerManager::getUpgradesForTower(string towerTypeID){
+  return allTowerUpgrades.at(towerTypeID);
 }
 
 /*
@@ -201,6 +333,11 @@ unordered_map<int, shared_ptr<TowerInterface>>& TowerManager::getTowersPlaced(){
  * @return TowerInterface: a pointer to the tower at the position on the grid
  */
 shared_ptr<TowerInterface> TowerManager::getTowerPlaced(int combinedRowCol){
+  if(towersPlaced.find(combinedRowCol) == towersPlaced.end()){
+    string noTowerID = textLoader -> getString("IDS_NT");
+    shared_ptr<TowerInterface> noTower = make_shared<NotATower>(noTowerID);
+    return noTower;
+  }
   return towersPlaced.at(combinedRowCol);
 }
 
@@ -210,6 +347,37 @@ shared_ptr<TowerInterface> TowerManager::getTowerPlaced(int combinedRowCol){
  * @return TowerInterface: the pointer to the tower at the position on the grid
  */
 shared_ptr<TowerInterface> TowerManager::getTowerPlaced(int row, int col){
+  if(towersPlaced.find(row*xDim+col) == towersPlaced.end()){
+    string noTowerID = textLoader -> getString("IDS_NT");
+    shared_ptr<TowerInterface> noTower = make_shared<NotATower>(noTowerID);
+    return noTower;
+  }
+  return towersPlaced.at(row*xDim + col);
+}
+
+/*
+ * Get the obstacle at the position as if it were a vector (used to grab
+ * whatever is available at a position for purchase)
+ * @param row: the row index for the obstacle
+ * @parm col: the col index for the obstacle
+ */
+vector<shared_ptr<TowerInterface>>& TowerManager::getObstacleAsVector(int row, int col){
+  shared_ptr<TowerInterface> obstacle = getObstacle( row,  col);
+  string type = obstacle -> getType();
+  assert(allTowerUpgrades.find(type) != allTowerUpgrades.end());
+  return allTowerUpgrades.at(type);
+}
+
+/*
+ * @param row: the row index for the obstacle
+ * @parm col: the col index for the obstacle
+ */
+shared_ptr<TowerInterface> TowerManager::getObstacle(int row, int col){
+  if(towersPlaced.find(row*xDim+col) == towersPlaced.end()){
+    string noTowerID = textLoader -> getString("IDS_NT");
+    shared_ptr<TowerInterface> noTower = make_shared<NotATower>(noTowerID);
+    return noTower;
+  }
   return towersPlaced.at(row*xDim + col);
 }
 
@@ -236,6 +404,35 @@ void TowerManager::handleTowerCreation(const EventInterface& event){
   //get the type of the tower (an index corresponding to a type in towersToChoose)
   string towerTypeID = tcEventData -> towerTypeID;
 
+  //decode the ID so that it is a row and column
+  int row = towerPosID / yDim;
+  int col = towerPosID % xDim;
+
+  //and add it
+  addTower(towerTypeID, row, col);
+}
+
+/*
+ * Add obstacles to the map where there are some
+ * @param allObstaclesToPlace: all unmade obstacles
+ * sent from GameLogic and BoardManager
+ */
+void TowerManager::addObstacles(unordered_map<int, intPair>& allObstaclesToPlace){
+  for(auto it = allObstaclesToPlace.begin(); it != allObstaclesToPlace.end(); ++it){
+      //grab the number stored at the position to indicate the type of obstacle
+      int typeNum = (*it).first;
+      //grab the string identifier of this type of obstacle
+      string obstacleType = textLoader -> getString(std::to_string(typeNum));
+      //the obstacle we will place
+      shared_ptr<TowerInterface> obstacle;
+
+      obstacle = copyOfTowerType(obstacleType);
+      //grab the position of the obstacle
+      int row = ((*it).second).first;
+      int col = ((*it).second).second;
+
+      towersPlaced.insert({row*xDim+col, obstacle});
+  }
 }
 
 
@@ -243,23 +440,114 @@ void TowerManager::handleTowerCreation(const EventInterface& event){
  * @param type: the index in the vector of tower types that
  * @param combinedRowCol: the posiition on the grid translated as the towers id
  */
-void TowerManager::addTower(int type, int combinedRowCol){
+void TowerManager::addTower(string type, int combinedRowCol){
+  //an tower of the desired type
+  shared_ptr<TowerInterface> aTower = copyOfTowerType(type);
 
+  //remove the current tower position if there is one
+  removeTower(combinedRowCol);
+  //insert the new tower
+  towersPlaced.insert({combinedRowCol, aTower});
 }
 /*
  * @param type: the index in the vector of tower types that
  * @param row: the posiition on the grid in the row
  * @param col: the position on the grid in the col
  */
-void TowerManager::addTower(int type, int row, int col){
+void TowerManager::addTower(string type, int row, int col){
 
+  //an tower of the desired type
+  shared_ptr<TowerInterface> aTower = copyOfTowerType(type);
+
+  //remove the current tower position if there is one
+  removeTower(row, col);
+  //insert the new tower
+  towersPlaced.insert({row*xDim+col, aTower});
 }
 
+/*
+ * @return a shared pointer to a copy of the tower type specified
+ */
+shared_ptr<TowerInterface> TowerManager::copyOfTowerType(string type){
+  shared_ptr<TowerInterface> retTower;
+
+  //get the maximum number of melee units allowed per melee tower
+  int maxMeleeUnits = textLoader -> getConstant(string("IDS_MT_MU"));
+
+  if(type == textLoader->getString(string("IDS_CPT"))){
+    retTower = make_shared<CheesePizza>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_PPT"))){
+    retTower = make_shared<PepperoniPizza>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_DDT"))){
+    retTower = make_shared<DeepDish>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_MLT"))){
+    retTower = make_shared<MeatLovers>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_MMMT"))){
+    retTower = make_shared<MiniMMS>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_NMMT"))){
+    retTower = make_shared<NormalMMS>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_PBMMT"))){
+    retTower = make_shared<PeanutButterMMS>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_SOT"))){
+    retTower = make_shared<PeanutMMS>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_EDT"))){
+    retTower = make_shared<EnergyDrink>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_GT"))){
+    retTower = make_shared<Gravy>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_SLT"))){
+    retTower = make_shared<Slushie>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_NFT"))){
+    retTower = make_shared<NormalFry>(textLoader, type, maxMeleeUnits);
+  }
+  else if(type == textLoader->getString(string("IDS_CFT"))){
+    retTower = make_shared<CrinkleFry>(textLoader, type, maxMeleeUnits);
+  }
+  else if(type == textLoader->getString(string("IDS_SFT"))){
+    retTower = make_shared<SpicyFry>(textLoader, type, maxMeleeUnits);
+  }
+  else if(type == textLoader->getString(string("IDS_WFT"))){
+    retTower = make_shared<WaffleFry>(textLoader, type, maxMeleeUnits);
+  }
+  else if(type == textLoader->getString(string("IDS_CTO"))){
+    retTower = make_shared<CafeteriaTable>(textLoader, type);
+  }
+  else if(type == textLoader->getString(string("IDS_TCO"))){
+    retTower = make_shared<TrashCan>(textLoader, type);
+  }
+  else
+  return retTower;
+}
+
+/*
+ * Remove a tower from the specified position
+ * @param combinedRowCol: the row*maxCol+col
+ */
 void TowerManager::removeTower(int combinedRowCol){
-
+  if(towersPlaced.find(combinedRowCol) != towersPlaced.end()){
+    towersPlaced.erase(combinedRowCol);
+  }
 }
+/*
+ * Remove a tower from the specified position
+ * @param row: row index
+ * @param col: the col index
+ */
 void TowerManager::removeTower(int row, int col){
-
+  int combinedRowCol = row * xDim + col;
+  if(towersPlaced.find(combinedRowCol) != towersPlaced.end()){
+    towersPlaced.erase(combinedRowCol);
+  }
 }
 
 void TowerManager::upgradeTower(int combinedRowCol){